[ai-form-recognizer] 5.0.0 changes (#26711)
This PR makes several changes related to the impending 5.0.0 release of
Document Intelligence (f.k.a. Form Recognizer).

- The major version revision signals a breaking change in which
key-value pairs and extracted languages will not be returned unless a
special `feature` is provided when making the analysis request.
- Support for explicit API version was removed. This functionality never
worked very well in TypeScript, and the major version revision gives us
an opportunity to remove it and avoid its inherent complexity until a
better solution is available.

This also addresses some issues with the way that training data content
sources were represented and named. The new representation is stricter
about providing only one of the two possible source type fields:
`azureBlobSource` or `azureBlobFileListSource`.

---------

Co-authored-by: Will Temple <will@wtemple.net>